* Npcs will respond only to player events * New random function, should work in windows and linux * Added mutex in Spawn::checkSpawn() * Fixed compiler warnings and move some functions to headers

+++ otserv/trunk/protocol79.cpp
@@ -100,7 +100,7 @@ void Protocol79::setPlayer(Player* p)
 
 void Protocol79::sleepTillMove()
 {
-	long long delay = player->getSleepTicks();
+	int64_t delay = player->getSleepTicks();
 	if(delay > 0 ){       
 #if __DEBUG__     
 		std::cout << "Delaying "<< player->getName() << " --- " << delay << std::endl;
